UPDATE_Updates-System_VBS_Windows7-Server2008
A method for forcing the OS to refresh certain system settings by re-reading the relevant registry keys. In particular for colour schemes or desktop backgrounds. So it meant that you could change e.g. the Windows desktop wallpaper by merging a .reg file into the registry and then executing UpdatePerUserSystemParameters from user32.dll to make it read and apply the changed entries.